Use double underscores as property name separator
Previously, if a property name was a prefix to another,
e.g. substrate and substrate_temperature, the form validation
would fail as _temperature would be considered information on
substrate instead of being recognized as an independent property.

By using double underscores as property name separators and
banning the use of double underscores in property names, this
commit fixes this issue.