Improve docker image size by using multi-stage build
Currently the image is 1.3 GB, this change will bring it down to 992 MB,
saving bandwidth and disk space to all.

The venv is built on an image and copied and used on another. This saves
us from installing gcc and git in the final image.

Some env vars were added:
DEBIAN_FRONTEND=noninteractive as is usual in apt scripts
PYTHONDONTWRITEBYTECODE because it's useless in a container
PYTHONUNBUFFERED don't try to buffer output
PIP_DISABLE_PIP_VERSION_CHECK not sure if useful here but doesn't hurt