Fix pylint errors for redefined functions
Starting with pylint 4.0, pylint's check for redefined functions also raises
errors for functions with a leading underscore, including the name "_". The
object search uses a lot of anonymous helper functions which are registered
using a decorator and otherwise not used, so having them anonymous is not
harmful and having the name _ redefined does not cause issues, so the error
has been disabled for that file.